Create readme.md
As practice will tell you, the English rules for comma placement are complex, frustrating, and often ambiguous. Many people, even the English, will, in practice,
ignore them, and, apply custom rules, or, no rules, at all.
Doctor Comma Sprinkler solved this issue by developing a set of rules that sprinkles
commas in a sentence with no ambiguity and little simplicity. In this problem you
will help Dr. Sprinkler by producing an algorithm to automatically apply her rules.
Dr. Sprinkler’s rules for adding commas to an existing piece of text are as follows:
1. If a word anywhere in the text is preceded by a comma, find all occurrences of that word in the text,
and put a comma before each of those occurrences, except in the case where such an occurrence
is the first word of a sentence or already preceded by a comma.
2. If a word anywhere in the text is succeeded by a comma, find all occurrences of that word in the
text, and put a comma after each of those occurrences, except in the case where such an occurrence
is the last word of a sentence or already succeeded by a comma.
3. Apply rules 1 and 2 repeatedly until no new commas can be added using either of them.